+++ grid2op/Chronics/__init__.py
@@ -8,16 +8,18 @@ __all__ = [
     "GridStateFromFileWithForecasts",
     "GridStateFromFileWithForecastsWithMaintenance",
     "GridStateFromFileWithForecastsWithoutMaintenance",
-    "ReadPypowNetData"
+    "ReadPypowNetData",
+    "FromNPY"
 ]
 
-from grid2op.Chronics.ChronicsHandler import ChronicsHandler
-from grid2op.Chronics.ChangeNothing import ChangeNothing
-from grid2op.Chronics.GridValue import GridValue
-from grid2op.Chronics.GridStateFromFile import GridStateFromFile
-from grid2op.Chronics.GridStateFromFileWithForecasts import GridStateFromFileWithForecasts
-from grid2op.Chronics.MultiFolder import Multifolder
-from grid2op.Chronics.ReadPypowNetData import ReadPypowNetData
+from grid2op.Chronics.chronicsHandler import ChronicsHandler
+from grid2op.Chronics.changeNothing import ChangeNothing
+from grid2op.Chronics.gridValue import GridValue
+from grid2op.Chronics.gridStateFromFile import GridStateFromFile
+from grid2op.Chronics.gridStateFromFileWithForecasts import GridStateFromFileWithForecasts
+from grid2op.Chronics.multiFolder import Multifolder
+from grid2op.Chronics.readPypowNetData import ReadPypowNetData
 from grid2op.Chronics.GSFFWFWM import GridStateFromFileWithForecastsWithMaintenance
-from grid2op.Chronics.FromFileWithoutMaintenance import GridStateFromFileWithForecastsWithoutMaintenance
-from grid2op.Chronics.MultifolderWithCache import MultifolderWithCache
+from grid2op.Chronics.fromFileWithoutMaintenance import GridStateFromFileWithForecastsWithoutMaintenance
+from grid2op.Chronics.multifolderWithCache import MultifolderWithCache
+from grid2op.Chronics.fromNPY import FromNPY