+++ grid2op/Observation/CompleteObservation.py
@@ -123,7 +123,8 @@ class CompleteObservation(BaseObservation):
         "is_alarm_illegal", "time_since_last_alarm", "last_alarm", "attention_budget",
         "was_alarm_used_after_game_over",
         "_shunt_p", "_shunt_q", "_shunt_v", "_shunt_bus",  # starting from grid2op version 1.6.0
-        "current_step", "max_step"  # starting from grid2op version 1.6.4
+        "current_step", "max_step",  # starting from grid2op version 1.6.4
+        "delta_time"  # starting grid2op version 1.6.5
     ]
     attr_list_json = ["_thermal_limit",
                       "support_theta",
@@ -235,3 +236,5 @@ class CompleteObservation(BaseObservation):
                 self.time_since_last_alarm[:] = -1
             self.last_alarm[:] = env._attention_budget.last_successful_alarm_raised
             self.attention_budget[:] = env._attention_budget.current_budget
+            
+        self.delta_time = env.delta_time_seconds / 60